feat: 1096 review returns 2
