+++ packages/backend/src/usagers/agrements/agrements.mail.ts
@@ -1,0 +1,129 @@
+import config, { frontUsagersDomain, senderEmail } from "../../config";
+import sendTemplate from "../../helpers/mail";
+import AppError from "../../utils/error";
+import logger from "../../utils/logger";
+
+const log = logger(module.filename);
+
+const assistanceText = `Si vous avez besoin d’accompagnement, vous pouvez contacter notre <a href='${config.assistance.uriFaq}'>équipe support</a>`;
+
+export const AgrementMailUsagers = {
+  sendStatutAModifierMail: ({
+    email,
+    commentaire,
+    regionDreets,
+  }: {
+    email: string[];
+    commentaire: string | undefined;
+    regionDreets: string;
+  }) => {
+    log.i("sendStatutAModifierMail - In", { email });
+    if (!email) {
+      throw new AppError(
+        "Email manquant pour l'envoi du mail de demande de modification d'agrément",
+      );
+    }
+    const html = sendTemplate.getBody(
+      "PeroVAO - Demande d'agrément transmise",
+      [
+        {
+          p: [
+            "Bonjour,",
+            `Suite à la réception de votre demande de renouvellement d’agrément, la DREETS ${regionDreets} vous informe qu’un ou plusieurs éléments complémentaires sont nécessaires afin de poursuivre l’examen de votre dossier.`,
+            "<strong>Commentaire de l’agent instructeur :</strong>",
+            commentaire?.replace(/\n/g, "<br>"),
+            "Nous vous invitons à vous connecter au portail <strong>VAO</strong> afin de consulter le détail de cette demande et de transmettre les informations ou documents demandés :",
+            `<a href=${frontUsagersDomain}>${frontUsagersDomain}</a>`,
+            "Pour toute question ou précision concernant cette demande de compléments, merci d’utiliser la <strong>messagerie intégrée au portail VAO</strong>, accessible depuis votre page agrément.",
+            assistanceText,
+          ],
+          type: "p",
+        },
+      ],
+      `L'équipe du SI VAO<BR><a href=${frontUsagersDomain}>Portail VAO</a>`,
+    );
+    const params = {
+      from: senderEmail,
+      html,
+      replyTo: senderEmail,
+      subject:
+        "Portail VAO - Demande de compléments d’informations suite à votre demande de renouvellement d’agrément",
+      to: email,
+    };
+    log.d("sendStatutTransmisMail post email", { params });
+    return params;
+  },
+  sendStatutCompletudeMail: ({
+    email,
+    regionDreets,
+  }: {
+    email: string[];
+    regionDreets: string;
+  }) => {
+    log.i("sendStatutCompletudeMail - In", { email });
+    if (!email) {
+      throw new AppError(
+        "Email manquant pour l'envoi du mail la validation de la complétude de l'agrément",
+      );
+    }
+    const urlAgrement = frontUsagersDomain + "/mon-agrement";
+    const html = sendTemplate.getBody(
+      "Portail VAO – Confirmation de la complétude de votre dossier de renouvellement d’agrément",
+      [
+        {
+          p: [
+            "Bonjour,",
+            `Votre demande de renouvellement d’agrément a été examinée par la DREETS ${regionDreets}.`,
+            "Conformément à la réglementation, l’administration dispose d’un délai de 2 mois pour instruire votre dossier et rendre une décision (ce délai est géré par la DREETS).",
+            `À défaut de décision dans ce délai, votre agrément sera réputé accordé (règle du "silence vaut accord"). La décision implicite d’acceptation du dossier ne court qu'à compter du moment où l’ensemble des pièces sont fournies par l’OVA.`,
+            "Vous pouvez suivre l’avancement de votre demande à tout moment depuis votre espace sur le portail VAO :",
+            `<a href=${urlAgrement}>${urlAgrement}</a>`,
+          ],
+          type: "p",
+        },
+      ],
+      `L'équipe du SI VAO<BR><a href=${frontUsagersDomain}>Portail VAO</a>`,
+    );
+    const params = {
+      from: senderEmail,
+      html,
+      replyTo: senderEmail,
+      subject:
+        "Portail VAO – Confirmation de la complétude de votre dossier de renouvellement d’agrément",
+      to: email,
+    };
+    log.d("sendStatutTransmisMail post email", { params });
+    return params;
+  },
+  sendStatutTransmisMail: ({ email }: { email: string }) => {
+    log.i("sendStatutTransmisMail - In", { email });
+    if (!email) {
+      throw new AppError(
+        "Email manquant pour l'envoi du mail de transmission d'agrément",
+      );
+    }
+    const html = sendTemplate.getBody(
+      "VAO - Demande d'agrément transmise",
+      [
+        {
+          p: [
+            "Bonjour,",
+            "Votre demande de renouvellement d'agrément a bien été transmise.",
+            assistanceText,
+          ],
+          type: "p",
+        },
+      ],
+      `L'équipe du SI VAO<BR><a href=${frontUsagersDomain}>Portail VAO</a>`,
+    );
+    const params = {
+      from: senderEmail,
+      html,
+      replyTo: senderEmail,
+      subject: "VAO - Demande d'agrément transmise",
+      to: email,
+    };
+    log.d("sendStatutTransmisMail post email", { params });
+    return params;
+  },
+};
